+++ eb_core/ecs/systems.rs
@@ -1,32 +1,10 @@
-use specs::{System, SystemData, Entities, ReadStorage, WriteStorage, Join};
+use specs::{System, Entities, ReadStorage, WriteStorage, Join, Fetch};
 use super::components::*;
+use types::{Track, Tile};
+
 
 pub struct ComputerRiderThink;
 
-pub enum TileKind {
-    Grass,
-    Dirt,
-    Hump,
-    Boost,
-    RampTiny,
-    RampMedium,
-    RampLarge,
-}
-
-pub struct Tile {
-    /// what lane the tile appears in
-    pub lane: u8,
-    /// the x position of the left edge of the tile
-    pub offset: f32,
-    /// the right edge of the tile is offset + width
-    pub width: f32,
-    /// The kind of tile this is
-    pub kind: TileKind,
-}
-
-pub struct Track {
-    tiles: Vec<Vec<Tile>>,
-}
 
 /// I guess this will return a 2d vec, one tile slice per lane, using some
 /// window to limit the tile count.
@@ -35,13 +13,10 @@ fn upcoming_tiles(track: &Track, from: f32, to: f32) -> &[Vec<Tile>] {
 }
 
 impl<'a> System<'a> for ComputerRiderThink {
-    type SystemData = (WriteStorage<'a, Bike>, ReadStorage<'a, ComputerRider>);
+    type SystemData = (WriteStorage<'a, Bike>, ReadStorage<'a, ComputerRider>, Fetch<'a, Track>);
 
-    fn run(&mut self, (mut bikes, brains): Self::SystemData) {
+    fn run(&mut self, (mut bikes, brains, track): Self::SystemData) {
         // FIXME: need access to the other bike positions, and terrain info.
-
-        let track = Track { tiles: vec![] };
-
         // Snapshotting the bike data so we can refer to it as we visit each one.
         // Each bike will need to know where other nearby bikes are before deciding what to do.
         let others = (&bikes).join().map(|x| x.clone()).collect::<Vec<Bike>>();
