+++ eb_core/lib.rs
@@ -3,3 +3,4 @@ extern crate specs;
 pub use specs::{DispatcherBuilder, World};
 
 pub mod ecs;
+pub mod types;
